refactor(web): split use-api.ts into domain-scoped hook files
Split the 1044-line monolithic use-api.ts into 8 focused files:
- use-dashboard-api.ts (overview stats, exec dashboard)
- use-traces-api.ts (GraphQL trace queries)
- use-sessions-api.ts (session management)
- use-agents-api.ts (agent CRUD, archive, drafts, versions)
- use-admin-api.ts (users, auth, audit, telemetry)
- use-review-api.ts (review actions, bundle review)
- use-registry-api.ts (component CRUD, versions)
- use-insights-api.ts (reports, models, feedback)

The original use-api.ts is now a barrel re-export, so all existing
imports continue to work unchanged.
